feat: add soft generator method (#56)
* feat: add soft generator method

* feat: fixed position toolbar and glass efects in options
