+++ package.json
@@ -2,6 +2,9 @@
   "name": "oxidus",
   "version": "0.1.0",
   "private": true,
+  "engines": {
+    "node": "22.x"
+  },
   "scripts": {
     "dev": "next dev",
     "build": "next build",
