+++ src/components/header/index.tsx
@@ -7,7 +7,7 @@ import { Menu } from "./menu";
 export default function Header() {
   return (
     <header
-      className={`h-[80px] space-x-4 flex items-center justify-between w-full py-2 px-6 border-b-[1px] border-b-foreground/10`}
+      className={`h-[64px] space-x-4 flex items-center justify-between w-full py-2 px-6 border-b-[1px] border-b-foreground/10`}
     >
       <div className="flex items-center space-x-2 relative h-full">
         <TypographyH1 className="text-xl">🪄 Oxidus</TypographyH1>
@@ -18,10 +18,6 @@ export default function Header() {
       </div>
       <CurrentThemeIndicatorContainer />
       <div className="flex items-center space-x-2">
-        <Menu.Themes />
-        <Menu.SaveChanges />
-        <Menu.ColorPicker />
-        <Menu.ShowCode />
         <Menu.SocialMedia />
         <Menu.ToggleTheme />
       </div>
