+++ src/components/header/menu/Themes.tsx
@@ -17,9 +17,9 @@ export default function ThemesMenuItem() {
     <>
       <Popover>
         <PopoverTrigger asChild>
-          <Button variant={"ghost"}>Themes</Button>
+          <Button className="h-fit w-fit" variant={"ghost"}>My Themes</Button>
         </PopoverTrigger>
-        <PopoverContent className="space-y-1 w-[360px] max-h-[400px] overflow-y-auto">
+        <PopoverContent className="bg-popover/20 backdrop-blur-xl space-y-1 w-[360px] max-h-[400px] overflow-y-auto">
           <div className="flex items-center justify-between border-b border-border pb-1">
             <TypographyParagraph>My Themes</TypographyParagraph>
             <AddNewThemeButton />
